Move namespace for dependency analyzers
