+++ src/common/Microsoft.DotNet.UpgradeAssistant.Abstractions/Dependencies/IDependencyAnalysisState.cs
@@ -1,7 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Microsoft.DotNet.UpgradeAssistant.Packages
+namespace Microsoft.DotNet.UpgradeAssistant.Dependencies
 {
     public interface IDependencyAnalysisState
     {
